make Verify Identity fields optional (#78)
* make verify identity fields optional

rework the VerifyMemberIdentityForm so that exp(iration date) is properly optional. Fix bug in the clean method. Name the field 'exp' to match VMI, and give it a label for the interface. Update the view accordingly.

* treat errors as lists

in the create member modal flow, errors are lists, so display all list members (usu. only one item)

* create org with bootstrap

add bootstrap container/row/col to the create org page

* update tests